tctl resources: convert db
