+++ tool/tctl/common/resources/resource.go
@@ -33,7 +33,8 @@ import (
 // to the Handler format.
 func Handlers() map[string]Handler {
 	return map[string]Handler{
-		types.KindRole: roleHandler(),
+		types.KindRole:     roleHandler(),
+		types.KindDatabase: databaseHandler(),
 	}
 }
